change for -Werror

+++ WebServer/WebServer/request.c
@@ -383,7 +383,7 @@ parse_requestpath_username(char *path, char *username, int length)
 	int checkindex;	
 	if (strncmp(path, "/~", 2) == 0 && strncmp(path, "/~/", 3) != 0) {
 		path = path + 2;
-		for (checkindex = 0; checkindex != strlen(path); ++checkindex) {
+		for (checkindex = 0; checkindex != (int)strlen(path); ++checkindex) {
 			if (path[checkindex] == '/' || checkindex == length)
 				break;
 			username[checkindex] = path[checkindex];
